Public export of Nexus code:   - bf0af38039831d5801a126d071261764da44ec57 by Matthew Piggott <mpiggott@sonatype.com>
GitOrigin-RevId: bf0af38039831d5801a126d071261764da44ec57

+++ components/nexus-repository-content/src/main/java/org/sonatype/nexus/repository/content/search/table/SearchTableDAO.java
@@ -39,7 +39,7 @@ public interface SearchTableDAO
    * @param values optional values map for filter (required if filter is not null)
    * @return number of found components.
    */
-  int count(@Nullable @Param("filter") String filter, @Nullable @Param("filterParams") Map<String, String> values);
+  long count(@Nullable @Param("filter") String filter, @Nullable @Param("filterParams") Map<String, String> values);
 
   /**
    * Search components in the scope of one format.
